minor syntax changes

+++ src/controllers/routes/categories.controllers.js
@@ -4,9 +4,9 @@ export async function getCategories(req, res) {
     try {
         const allCategories = await Category.findAll();
 
-        if (allCategories.length === 0) return res.status(204).json({ 'msg': 'No categories have been found' });
+        if (allCategories.length === 0) return res.status(204).send({ 'msg': 'No categories have been found' });
 
-        return res.json(allCategories);
+        return res.send(allCategories);
     } catch (error) {
         console.error(error);
     }
@@ -18,9 +18,9 @@ export async function getCategoryById(req, res) {
     try {
         const foundCategory = await Category.findByPk(id, { include: ["products"] });
 
-        !foundCategory && res.status(204).json({ error: `There are no Categories with the id : ${id}` });
+        if (!foundCategory) return res.status(204).send({ error: `There are no Categories with the id : ${id}` });
 
-        return res.json(foundCategory);
+        return res.send(foundCategory);
 
     } catch (error) {
         console.error(error)
@@ -38,12 +38,12 @@ export async function postCategory(req, res) {
             }
         });
 
-        if (!name) return res.status(400).json({ error: "Must provide all the required fields" });
+        if (!name) return res.status(400).send({ error: "Must provide all the required fields" });
 
-        if (categoryWithSameName) return res.status(405).json({ error: `There's already a category with that name` });
+        if (categoryWithSameName) return res.status(405).send({ error: `There's already a category with that name` });
 
         const newCategory = await Category.create({ name });
-        return res.status(201).json(newCategory);
+        return res.status(201).send(newCategory);
     } catch (error) {
         console.error(error)
     }
@@ -55,12 +55,12 @@ export async function modifyCategory(req, res) {
     try {
         const foundCategory = await Category.findByPk(id);
 
-        !foundCategory && res.status(204).json({ error: `There are no Categories with the id : ${id}` });
+        if (!foundCategory) return res.status(204).send({ error: `There are no Categories with the id : ${id}` });
 
         await Category.update(body, { where: { id: id } });
         const updatedCategory = await Category.findByPk(id);
 
-        return res.json(updatedCategory);
+        return res.send(updatedCategory);
     } catch (error) {
         console.error(error)
     }
@@ -72,11 +72,11 @@ export async function deleteCategory(req, res) {
     try {
         const foundCategory = await Category.findByPk(id);
 
-        !foundCategory && res.status(204).json({ error: `There are no categories with the id : ${id}` });
+        if (!foundCategory) return res.status(204).send({ error: `There are no categories with the id : ${id}` });
 
         await Category.destroy({ where: { id: id } });
 
-        return res.status(200).json({ msg: `Category ${foundCategory.name} - id : ${id} deleted` });
+        return res.status(200).send({ msg: `Category ${foundCategory.name} - id : ${id} deleted` });
 
     } catch (err) {
         console.log(err)
